requirements.txt
Flask==2.1.1
pydantic==1.9.0
boto3==1.21.32
python-decouple==3.6
pre-commit==2.18.1
gunicorn==20.1.0
botocore==1.24.32
rollbar==0.16.2
blinker==1.4
